Refactor shared model
This commit just refactored shared model for admin.
