+++ BackEnd/Services/CategoryService.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using BackEnd.Interfaces;
 using BackEnd.Models;
-using Shared.Admin;
+using BackEnd.Models.ViewModels;
 using Shared.Clients;
 
 namespace BackEnd.Services
